pyproject.toml
[project]
name = "tele-convo"
version = "0.1.0"
description = "Telegram message scraper with SQLite storage and JSON-RPC WebSocket interface"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "telethon>=1.42.0",
    "websockets>=16.0",
    "python-dotenv>=1.2.1",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
